server/requirements.txt
fastapi
pydantic
uvicorn
tensorflow
numpy
matplotlib
langchain_core
langchain_groq
pillow
